+++ frontend/src/tables/elements/TableReloadButton.tsx
@@ -1,0 +1,21 @@
+import { Button } from "react-bootstrap";
+import type { createTableStore } from "../TableStoreFactory";
+import { ArrowRepeat } from "react-bootstrap-icons";
+
+interface Props {
+  useStore: ReturnType<typeof createTableStore>;
+}
+
+export default function TableReloadButton({ useStore }: Props) {
+  const tableStore = useStore();
+  return (
+    <Button
+      type="button"
+      variant=""
+      className="mx-3 fw-bold p-0"
+      onClick={() => tableStore.setNeedReload(true)}
+    >
+      <ArrowRepeat size="1.75rem" className="fw-bold"></ArrowRepeat>
+    </Button>
+  );
+}